Extract HGV scenario mapping into its own module to fix circular import
SCENARIO_MAP and active_scenario_map() lived in hgv_charging/__init__.py
but were needed by its own submodules (fill_tables, spatial_assignment,
etrago_integration) at import time, causing a circular import. Move
them into a new scenarios.py module that all four files import from.
