+++ src/egon/data/datasets/emobility/hgv_charging/fill_tables.py
@@ -17,7 +17,7 @@ import geopandas as gpd
 import pandas as pd
 
 from egon.data import config, db
-from egon.data.datasets.emobility.hgv_charging import active_scenario_map
+from egon.data.datasets.emobility.hgv_charging.scenarios import active_scenario_map
 
 
 def _input_dir() -> Path:
